+++ simpleSdlWrapper/imagen.h
@@ -4,7 +4,7 @@
 #define MAX_IMG 100
 
 int  irudiaKargatu(char *fileName);
-void  irudiaMugitu(int numImg, int x, int y);
+void  irudiaMugitu(int numImg, float x, float y);
 void irudiakMarraztu(void);
 void irudiaKendu(int id);
